# optimize test_disk_exhaustion to remove huge dummy file after use