<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary: Optimize the flow of test case test_disk_exhaustion
Fixes # (issue)

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [x] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

### Approach
#### What is the motivation for this PR?
The original implementation of this testcase is to exhaust(95% of available space) /tmp folder in a script running in background and then sleep 60s before deleting the huge dummy file. This is redundant to sleep 60s and may cause check failure when the /tmp is tmpfs and available space is huge(e.g. > 30GB). The fallocate >30GB operation against tmpfs may take > 10s and cause the check of used_after_test with uaed_before_test to fail.

#### How did you do it?
Run fallocate in foreground and remove the 60s sleep. Remove huge dummy file in the finally block
#### How did you verify/test it?
Verified the case passed with a testbed that has tmpfs /tmp with size > 30GB
#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->